Assets/Websockets/requirements.txt
pathlib
websockets
pyserial
websocket-server